Refactor(#48): config 위치 수정
